requirements.txt
altair==5.3.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==5.0.6
exceptiongroup==1.2.1
executing==2.0.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
stack-data==0.6.3
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
